reader_concurrency_semaphore: add eviction notification facility
Allow passing a notification functor when registering an inactive read.
This functor will be called when (and if) the read is evicted.
We want to move away from `inactive_read` and store flat mutation
readers in the semaphore directly, and this mechanism will allow users
who relied on `inactive_read::evict()` to notify them about the eviction
to move away from `inactive_read::evict()`.