reader_concurrency_semaphore: introduce ttl for inactive reads
We want to migrate all inactive read caching logic from the querier
cache, to the reader concurrency semaphore so its all in one place,
leaving only indexing to the former. One of the last remaining features
implemented in the querier cache is ttl. Add an equivalent feature to
the semaphore to allow the migration.