querier_cache: store readers in the reader concurrency semaphore directly
Currently the reader is stored in the querier and the inactive read
object that is registered with the reader concurrency semaphore only has
a pointer to it, which it uses to eviction. We want to retire the
inactive read class and require its users to store the readers directly.
To make this migration easier store the reader directly in the inactive
read instance when the querier is cached.
With this the `_entries` list storing the queriers no longer really
serves a purpose. It is not needed for eviction as that was outsourced
to the semaphore in previous patches. We can store the querier object
directly in the indexes. After this, the querier cache is not storing
the cached readers anymore, its only job is indexing and maintaining
relevant stats.