querier_cache: delegate expiry to the reader_concurrency_semaphore
Using the newly introduced ttl for inactive reads feature. Another step
towards moving all cacheing related functionality to the semaphore, so
its all in one place.
A side effect of this patch is that changing the entry ttl of the cache
doesn't apply it on current entries anymore. Currently only tests change
the entry ttl and they don't rely on it being applied to existing
entries, so this is fine.